+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Features/Changes
 * Compiler: try to preserve clorures ordering between ml and js
+* Compiler: no longer rely on IIFE for scoping variable inside loops
 
 ## Bug fixes
 * Compiler: js-parser now accept all the line terminators defined in the spec