+++ CHANGES.md
@@ -1,4 +1,4 @@
-# 5.5.2 (2023-12-01) - Lille
+# dev (2023-??-??) - Lille
 
 ## Features/Changes
 * Compiler: try to preserve clorures ordering between ml and js
@@ -7,6 +7,7 @@
 * Compiler: js-parser now accept all the line terminators defined in the spec
 * Compiler: js-parser: fix support for LHS assignment target
 * Compiler: js-parser: fix parser of default export
+* Compiler: js-parser: allow 'as' as ident
 
 # 5.5.2 (2023-12-01) - Lille